compile data code to 2 files
