+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack --mode production --watch",
-    "start": "node db/generateData.js"
+    "start": "node db/schema.js"
   },
   "lint-staged": {
     "*.js": "eslint",
